exception but bugs

+++ include/kernel.h
@@ -8,8 +8,8 @@
 #include <include/string.h>
 
 typedef unsigned char      u8;
-typedef unsigned short     u16;
-typedef unsigned int       u32;
+typedef unsigned short int u16;
+typedef unsigned long int  u32;
 typedef unsigned long long u64;
 typedef char               s8;
 typedef short              s16;
